+++ README.md
@@ -1,4 +1,4 @@
-# otus spring homework 11
-homework 11 for Otus springFramework course
+# otus spring homework 12
+homework 12 for Otus springFramework course
 
 Spring authentication
