--- src/main/java/ru/ezikvice/springotus/homework11/Homework8Application.java
@@ -1,12 +1,0 @@
-package ru.ezikvice.springotus.homework11;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Homework8Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Homework8Application.class);
-    }
-}
