--- src/main/java/ru/ezikvice/springotus/homework11/domain/Comment.java
@@ -1,19 +1,0 @@
-package ru.ezikvice.springotus.homework11.domain;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-
-@Data
-@NoArgsConstructor
-public class Comment {
-
-    @Id
-    private String id;
-
-    private String text;
-
-    public Comment(String text) {
-        this.text = text;
-    }
-}
