--- src/main/java/ru/ezikvice/springotus/homework11/service/AuthorService.java
@@ -1,22 +1,0 @@
-package ru.ezikvice.springotus.homework11.service;
-
-import ru.ezikvice.springotus.homework11.domain.Author;
-
-import java.util.List;
-import java.util.Set;
-
-public interface AuthorService {
-
-    long count();
-
-    Author save(Author author);
-
-    void delete(Author author);
-
-    Author findById(String id);
-
-    List<Author> findAll();
-
-    Set<Author> findByName(String name);
-
-}
